fix: getting rid of d.ts linting errors, exporting properly
At least I think this exports it properly; seems to work in apps!
