+++ Makefile
@@ -13,7 +13,7 @@ cat_docs_command = cat ./docs/_API-header.md ./docs/_API-body.md > ./docs/API.md
 build:
 	rm -rf ./dist
 	tsc
-	cp index.d.ts dist/index.d.ts
+	cp frontend-base.d.ts dist/frontend-base.d.ts
 	mkdir -p ./scss/header/studio-header
 
 	cp shell/index.scss dist/shell/index.scss
